Split processing from avh processing file that was relevant for profiles Updated config to reflect change
