+++ dataSources/ala/avh/processing.py
@@ -1,9 +1,7 @@
 import pandas as pd
 import requests
-import json
-import math
 from pathlib import Path
-import lib.dataframeFuncs as dff
+
 
 def build(outputFilePath: Path) -> None:
     baseURL = "https://biocache-ws.ala.org.au/ws/occurrences/search?q=*%3A*&disableAllQualityFilters=true&qualityProfile=AVH&fq=type_status%3A*&fq=country%3A%22Australia%22&qc=data_hub_uid%3Adh9"
@@ -14,7 +12,7 @@ def build(outputFilePath: Path) -> None:
     jsData = rawData.json()
 
     records = jsData["totalRecords"]
-    totalCalls = math.ceil(records / readSize)
+    totalCalls = (records / readSize).__ceil__()
 
     occurrences = []
     for call in range(totalCalls):
@@ -26,32 +24,3 @@ def build(outputFilePath: Path) -> None:
 
     df = pd.DataFrame.from_records(occurrences)
     df.to_csv(outputFilePath, index=False)
-
-def collect(outputDir: Path, profile: str, tokenFilePath: Path) -> None:
-    with open(tokenFilePath) as fp:
-        token = json.load(fp)
-
-    bearerToken = token["access_token"]
-    baseURL = "https://api.ala.org.au/profiles"
-    endpoint = f"/api/opus/{profile}/profile?pageSize=1000"
-    response = requests.get(baseURL + endpoint, headers={"Authorization": f"Bearer {bearerToken}"})
-    data = response.json()
-
-    if "message" in data and "not authorized" in data["message"]:
-        print("Failed to authorize, please make sure bearer token is valid.")
-        return
-    
-    print(f"Accessing profile: {profile}")
-
-    records = []
-    for idx, entry in enumerate(data, start=1):
-        uuid = entry["uuid"]
-        print(f"At record: {idx}", end="\r")
-
-        response = requests.get(baseURL + f"/api/opus/{profile}/profile/{uuid}", headers={"Authorization": f"Bearer {bearerToken}"})
-        records.append(response.json())
-    print()
-
-    df = pd.DataFrame.from_records(records)
-    df = dff.removeSpaces(df)
-    df.to_csv(outputDir / f"{profile}.csv", index=False)
