+++ dataSources/ala/profiles/config.json
@@ -8,7 +8,7 @@
         "weeds-australia": {}
     },
     "download": {
-        "path": "sourceProcessing/ala.py",
+        "path": "./processing.py",
         "function": "collect",
         "args": [
             "{OUTDIR}",
